old_requirements.txt
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dlib @ file:///D:/bld/dlib-split_1723251352733/work
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
hdbscan==0.8.40
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.41.1
lpips==0.1.4
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
ninja==1.13.0
numba==0.58.1
numpy @ file:///D:/bld/numpy_1687808486521/work
opencv-python==4.12.0.88
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pynndescent==0.6.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.3
umap-learn==0.5.7
zipp==3.20.2
